+++ frontend/src/routes/Infrastructure/InfraEnvironments/InfraEnvironmentsPage.test.tsx
@@ -1,12 +1,7 @@
 /* Copyright Contributors to the Open Cluster Management project */
 import { render } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import {
-  AgentK8sResource,
-  AgentServiceConfigK8sResource,
-  InfraEnvK8sResource,
-  SecretK8sResource,
-} from '@openshift-assisted/ui-lib/cim'
+import { AgentK8sResource, AgentServiceConfigK8sResource, InfraEnvK8sResource } from '@openshift-assisted/ui-lib/cim'
 import { MemoryRouter, Route, Routes } from 'react-router-dom-v5-compat'
 import { RecoilRoot } from 'recoil'
 
@@ -21,48 +16,7 @@ import InfraEnvironmentsPage, {
   isDeleteDisabled,
   isPullSecretReused,
 } from './InfraEnvironmentsPage'
-
-export const infraEnvName = 'infra-env-name'
-
-export const mockInfraEnv1: InfraEnvK8sResource = {
-  apiVersion: 'agent-install.openshift.io/v1beta1',
-  kind: 'InfraEnv',
-  metadata: {
-    labels: {
-      'agentclusterinstalls.extensions.hive.openshift.io/location': 'brno',
-      networkType: 'dhcp',
-    },
-    name: infraEnvName,
-    namespace: infraEnvName,
-  },
-  spec: {
-    agentLabels: {
-      'agentclusterinstalls.extensions.hive.openshift.io/location': 'brno',
-    },
-    pullSecretRef: {
-      name: `pullsecret-${infraEnvName}`,
-    },
-    cpuArchitecture: 'x86_64',
-  },
-  status: {
-    agentLabelSelector: {
-      matchLabels: {
-        'infraenvs.agent-install.openshift.io': infraEnvName,
-      },
-    },
-    conditions: [
-      {
-        lastTransitionTime: '2021-10-04T11:26:37Z',
-        message: 'Image has been created',
-        reason: 'ImageCreated',
-        status: 'True',
-        type: 'ImageCreated',
-      },
-    ],
-    createdTime: '2021-11-10T13:00:00Z',
-    isoDownloadURL: 'https://my.funny.download.url',
-  },
-}
+import { infraEnvName, mockInfraEnv1 } from '../../../test-helpers/infraEnvName'
 
 const mockAgent1: AgentK8sResource = {
   apiVersion: 'agent-install.openshift.io/v1beta1',
@@ -96,21 +50,6 @@ const mockAgent2: AgentK8sResource = {
     hostname: 'host',
     role: 'auto-assign',
   },
-}
-
-export const mockPullSecret: SecretK8sResource = {
-  apiVersion: 'v1',
-  kind: 'Secret',
-  metadata: {
-    name: `pullsecret-${infraEnvName}`,
-    namespace: infraEnvName,
-    labels: { 'cluster.open-cluster-management.io/backup': 'cluster' },
-  },
-  data: {
-    '.dockerconfigjson':
-      'eyJhdXRocyI6eyJjbG91ZC5vcGVuc2hpZnQuY29tIjp7ImF1dGgiOiJiM0JsYlNLSVBQRUQiLCJlbWFpbCI6Im15QGVtYWlsLnNvbWV3aGVyZS5jb20ifX19',
-  },
-  type: 'kubernetes.io/dockerconfigjson',
 }
 
 const mockInfraEnvironments: InfraEnvK8sResource[] = [mockInfraEnv1]
